fixed bug which didn't allow user to skip uploading an image

+++ routes/add.js
@@ -5,7 +5,7 @@ var fs = require('fs');
 mongoose.createConnection(process.env.MONGO_URL);
 
 var Robot = require('../schemas/robot')
-/* GET add page. */
+    /* GET add page. */
 router.get('/', function(req, res) {
     res.render('add');
 });
@@ -13,15 +13,17 @@ router.get('/', function(req, res) {
 /* POST add page. */
 router.post('/', function(req, res) {
     console.log('fssfds', req.body.other)
-    fs.readFileSync('uploads/' + req.files.image.name)
     var robot = new Robot({
         teamNumber: req.body.teamNumber,
         abilities: req.body.abilities,
         other: req.body.other
     });
-    robot.img.data = fs.readFileSync('uploads/' + req.files.image.name);
-    robot.img.contentType = 'image/png';
-    fs.unlinkSync('uploads/' + req.files.image.name)
+    if (req.files.image) {
+        fs.readFileSync('uploads/' + req.files.image.name)
+        robot.img.data = fs.readFileSync('uploads/' + req.files.image.name);
+        robot.img.contentType = 'image/png';
+        fs.unlinkSync('uploads/' + req.files.image.name)
+    }
     robot.save();
     res.send('Done. It might take a while for it to show up in /robots')
